refactor(persona): move useClickOutside hook to shared
