+++ mods/persona/shared/hooks/useClickOutside.ts
@@ -1,0 +1,19 @@
+import { useEffect, type RefObject } from "react";
+
+/**
+ * Calls `onClickOutside` when a mousedown event occurs outside the referenced element.
+ */
+export function useClickOutside(
+  ref: RefObject<HTMLElement | null>,
+  onClickOutside: () => void,
+): void {
+  useEffect(() => {
+    function handleMouseDown(e: MouseEvent) {
+      if (ref.current && !ref.current.contains(e.target as Node)) {
+        onClickOutside();
+      }
+    }
+    document.addEventListener("mousedown", handleMouseDown);
+    return () => document.removeEventListener("mousedown", handleMouseDown);
+  }, [ref, onClickOutside]);
+}
